src/requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.2
openai==1.6.1
chromadb==0.4.22
python-dotenv==1.0.0
tiktoken==0.5.2